requirements.txt
# EcoVation Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
tavily-python>=0.3.0
Pillow>=10.0.0
python-multipart>=0.0.6
jinja2>=3.1.0
